exercise 4.15-4.23

+++ controller/post.controller.js
@@ -1,15 +1,25 @@
 const Blog = require('../models/post.model')
+const User = require('../models/users.model')
+const requiredLogin = require('../middle/requireLogin')
 const PostRoute = require('express').Router()
 const logger = require('../util/logger')
+require('dotenv').config()
 
-PostRoute.post('/', async (req, res, next) => {
+PostRoute.post('/', requiredLogin, async (req, res, next) => {
     const { title, author, url, likes } = req.body
+    const { user } = req;
     try {
-        let post = new Blog({
-            title, author,
-            url, likes
+        let post = await new Blog({
+            title,
+            author,
+            url,
+            likes,
+            user: user?._id
         })
-        await post.save()
+        logger.info(req.user)
+        const savedBlog = await post.save()
+        user.blog = user.blog.concat(savedBlog._id)
+        await user.save()
         res.status(201).json(post)
     } catch(err) {
         logger.error(err)
@@ -38,11 +48,11 @@ PostRoute.get('/:postId', async (req, res, next) => {
     }
 })
 
-PostRoute.put('/:postId', async (req, res, next) => {
+PostRoute.put('/:postId', requiredLogin, async (req, res, next) => {
     const { postId } = req.params
     const { title, author, url, likes } = req.body
     try {
-        const response = await Blog.findByIdAndUpdate(postId, {title, author, url, likes})
+        const response = await Blog.findByIdAndUpdate(postId, { title, author, url, likes })
         res.status(200).json(response)
     } catch(err) {
         logger.error(err)
@@ -50,11 +60,16 @@ PostRoute.put('/:postId', async (req, res, next) => {
     }
 })
 
-PostRoute.delete('/:postId', async (req, res, next) => {
+PostRoute.delete('/:postId', requiredLogin, async (req, res, next) => {
     const { postId } = req.params
+    const blog = await Blog.findById(postId)
     try {
-        await Blog.findByIdAndDelete(postId)
-        res.status(204).end()
+        if(req?.user?._id.toString() === blog?.user?._id?.toString()) {
+            await Blog.findByIdAndDelete(postId)
+            res.status(204).end()
+        } else {
+            return res.status(401).end()
+        }
     } catch(err) {
         logger.error(err)
         next(err)
